Add table assignment functionality to reservation system
This commit introduces the ability to assign a table to a reservation. New methods were added to the ReservationService and ReservationRepository classes to fetch available tables for a given time slot and assign a chosen table to a reservation. These changes also involved updates to the AssignTableController and assigntable.html to allow user interaction with the new functionality.